refactor: profile -> home으로 변경
