+++ frontend/app/src/main/java/com/markoala/tomoandroid/navigation/AppNavHost.kt
@@ -4,13 +4,12 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-
+import com.markoala.tomoandroid.ui.home.HomeScreen
 import com.markoala.tomoandroid.ui.login.LoginScreen
-import com.markoala.tomoandroid.ui.profile.ProfileScreen
 
 sealed class Screen(val route: String) {
     object Login : Screen("login")
-    object Profile : Screen("profile")
+    object Profile : Screen("home")
 }
 
 @Composable
@@ -23,7 +22,7 @@ fun AppNavHost(navController: NavHostController, isSignedIn: Boolean) {
             LoginScreen(navController)
         }
         composable(Screen.Profile.route) {
-            ProfileScreen(onSignOut = {
+            HomeScreen(onSignOut = {
                 navController.navigate(Screen.Login.route) {
                     popUpTo(Screen.Profile.route) { inclusive = true }
                 }
